chore(napi): zero clippy warnings in the surface, and lower the size ratchet
- Two `u64 -> u64` casts left over from the DTO collapse.
- `napi_engine.rs` is 2,790 lines from 5,915, so it comes off
  `prod_files_over_cap` (4 -> 3). The ratchet only ever moves down.
- README: one table saying which generator writes which types, and the one
  consequence of the JSON transport a caller can observe — an absent optional
  is an absent key, not `null`.

Verified: `pnpm generate:types` twice in a row and a fresh addon build both
produce a zero-line diff, so CI's "Generated files are current" step has a
stable signal for both locations it already covers (`crates/spaghetti-napi`
for `index.d.ts`, `packages/sdk/src/generated` for the ts-rs bindings).

clippy 0 warnings in engine/*, napi_engine.rs, napi_catalog.rs, lib.rs;
148 remain, all in `adapter/*` (86) and `runtime_semantic_reducer.rs` (62),
which L5 owns. `cargo check --features legacy-oracle` clean. validate-all 9/9.

+++ crates/spaghetti-napi/src/napi_engine.rs
@@ -1672,8 +1672,8 @@ impl Task for StartConfiguredObservationTask {
             .start_configured_observation_cancellable(configured, self.cancellation.clone())
             .map_err(napi_error)?;
         encode_json(&CatalogStartup {
-            catalog_projects: outcome.catalog_projects as u64,
-            catalog_sessions: outcome.catalog_sessions as u64,
+            catalog_projects: outcome.catalog_projects,
+            catalog_sessions: outcome.catalog_sessions,
             degraded_sources: outcome.degraded_sources,
             supervisors_started: u32::try_from(outcome.supervisors_started).unwrap_or(u32::MAX),
             history_background: outcome.history_background,
